+++ R/continue_funs.R
@@ -122,6 +122,7 @@ kmbayes_continue <- function(fit, ...){
 #' @param ... arguments to \code{\link[bkmrhat]{kmbayes_continue}}
 #'
 #' @return a `bkmrfit.list` object, which is just a list of `bkmrfit` objects similar to \code{\link[bkmrhat]{kmbayes_parallel}}
+#' @importFrom future value
 #' @export
 #' @seealso \code{\link[bkmrhat]{kmbayes_parallel}}
 #' @import future
@@ -133,8 +134,8 @@ kmbayes_continue <- function(fit, ...){
 #' Z <- dat$Z
 #' X <- dat$X
 #' \dontrun{
-#' Sys.setenv(R_FUTURE_SUPPORTSMULTICORE_UNSTABLE="quiet")
-#' future::plan(strategy = future::multiprocess, workers=2)
+#' 
+#' future::plan(strategy = future::multisession, workers=2)
 #' fitty1p = kmbayes_parallel(nchains=2, y=y,Z=Z,X=X)
 #'
 #' fitty2p = kmbayes_parallel_continue(fitty1p, iter=3000)
@@ -151,7 +152,7 @@ kmbayes_parallel_continue <- function(fitkm.list, ...) {
       kmbayes_continue(fitkm.list[[ii]], ...)
     }, seed=ss[ii])
   }
-  res <- values(ff)
+  res <- value(ff)
   class(res) <- c("bkmrfit.list", class(res))
   res
 }